Added rescan() method to the detect engine.
Added support for overriding the eslint-config-axway rules with a project specific .eslintrc file.

Fixed the plugin system to explicitly declare if the process is the parent instead of checking if the IPC channel exists, which it now always does.

Added support for detecting Xcodes and watching them in the iOS plugin.

Cleaned up some styling in the Windows plugin.

Updated npm deps.