[DAEMON-146] Fixed detect engine to not scan an already discovered path.
Added the ability to the detect engine to specify a relative prefix to the exe to be applied after resolving it.

Fixed bug in detect engine where the searchPaths was not being properly exposed.

Added the /appcd/health service endpoint and wired up pulling the agent data from the core process and all running external plugins.

More cleanup of the appcd config command.

Added the health info to the appcd dump command.

Fixed selecting the correct default Android SDK in the Android plugin.

Added docs for the appcd dump command.